+++ notify-impl/src/main/java/com/dwarfeng/notify/impl/bean/entity/HibernateTopic.java
@@ -1,0 +1,73 @@
+package com.dwarfeng.notify.impl.bean.entity;
+
+import com.dwarfeng.notify.sdk.util.Constraints;
+import com.dwarfeng.subgrade.sdk.bean.key.HibernateStringIdKey;
+import com.dwarfeng.subgrade.stack.bean.Bean;
+
+import javax.persistence.*;
+import java.util.Optional;
+
+@Entity
+@IdClass(HibernateStringIdKey.class)
+@Table(name = "tbl_topic")
+public class HibernateTopic implements Bean {
+
+    private static final long serialVersionUID = -3414846759646764018L;
+
+    // -----------------------------------------------------------主键-----------------------------------------------------------
+    @Id
+    @Column(name = "id", nullable = false, unique = true)
+    private String stringId;
+
+    // -----------------------------------------------------------主属性字段-----------------------------------------------------------
+    @Column(name = "label", length = Constraints.LENGTH_LABEL)
+    private String label;
+
+    @Column(name = "remark", length = Constraints.LENGTH_REMARK)
+    private String remark;
+
+    public HibernateTopic() {
+    }
+
+    // -----------------------------------------------------------映射用 getter&setter-----------------------------------------------------------
+    public HibernateStringIdKey getKey() {
+        return Optional.ofNullable(stringId).map(HibernateStringIdKey::new).orElse(null);
+    }
+
+    public void setKey(HibernateStringIdKey key) {
+        this.stringId = Optional.ofNullable(key).map(HibernateStringIdKey::getStringId).orElse(null);
+    }
+
+    // -----------------------------------------------------------常规 getter&setter-----------------------------------------------------------
+    public String getStringId() {
+        return stringId;
+    }
+
+    public void setStringId(String stringId) {
+        this.stringId = stringId;
+    }
+
+    public String getLabel() {
+        return label;
+    }
+
+    public void setLabel(String label) {
+        this.label = label;
+    }
+
+    public String getRemark() {
+        return remark;
+    }
+
+    public void setRemark(String remark) {
+        this.remark = remark;
+    }
+
+    @Override
+    public String toString() {
+        return getClass().getSimpleName() + "(" +
+                "stringId = " + stringId + ", " +
+                "label = " + label + ", " +
+                "remark = " + remark + ")";
+    }
+}