+++ notify-impl/src/main/java/com/dwarfeng/notify/impl/cache/TopicCacheImpl.java
@@ -1,0 +1,97 @@
+package com.dwarfeng.notify.impl.cache;
+
+import com.dwarfeng.notify.sdk.bean.entity.FastJsonTopic;
+import com.dwarfeng.notify.stack.bean.entity.Topic;
+import com.dwarfeng.notify.stack.cache.TopicCache;
+import com.dwarfeng.subgrade.impl.cache.RedisBatchBaseCache;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
+import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+import com.dwarfeng.subgrade.stack.exception.CacheException;
+import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Repository
+public class TopicCacheImpl implements TopicCache {
+
+    private final RedisBatchBaseCache<StringIdKey, Topic, FastJsonTopic> topicBatchBaseDelegate;
+
+    public TopicCacheImpl(
+            RedisBatchBaseCache<StringIdKey, Topic, FastJsonTopic> topicBatchBaseDelegate
+    ) {
+        this.topicBatchBaseDelegate = topicBatchBaseDelegate;
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean exists(StringIdKey key) throws CacheException {
+        return topicBatchBaseDelegate.exists(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public Topic get(StringIdKey key) throws CacheException {
+        return topicBatchBaseDelegate.get(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void push(Topic value, long timeout) throws CacheException {
+        topicBatchBaseDelegate.push(value, timeout);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void delete(StringIdKey key) throws CacheException {
+        topicBatchBaseDelegate.delete(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void clear() throws CacheException {
+        topicBatchBaseDelegate.clear();
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean allExists(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        return topicBatchBaseDelegate.allExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean nonExists(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        return topicBatchBaseDelegate.nonExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @SkipRecord
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public List<Topic> batchGet(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        return topicBatchBaseDelegate.batchGet(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchPush(@SkipRecord List<Topic> entities, long timeout) throws CacheException {
+        topicBatchBaseDelegate.batchPush(entities, timeout);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchDelete(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        topicBatchBaseDelegate.batchDelete(keys);
+    }
+}