+++ notify-stack/src/main/java/com/dwarfeng/notify/stack/bean/entity/Topic.java
@@ -1,0 +1,63 @@
+package com.dwarfeng.notify.stack.bean.entity;
+
+import com.dwarfeng.subgrade.stack.bean.entity.Entity;
+import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+
+/**
+ * 主题。
+ *
+ * @author DwArFeng
+ * @since 1.0.0
+ */
+public class Topic implements Entity<StringIdKey> {
+
+    private static final long serialVersionUID = -574921187295491197L;
+
+    private StringIdKey key;
+    private String label;
+    private String remark;
+
+    public Topic() {
+    }
+
+    public Topic(StringIdKey key, String label, String remark) {
+        this.key = key;
+        this.label = label;
+        this.remark = remark;
+    }
+
+    @Override
+    public StringIdKey getKey() {
+        return key;
+    }
+
+    @Override
+    public void setKey(StringIdKey key) {
+        this.key = key;
+    }
+
+    public String getLabel() {
+        return label;
+    }
+
+    public void setLabel(String label) {
+        this.label = label;
+    }
+
+    public String getRemark() {
+        return remark;
+    }
+
+    public void setRemark(String remark) {
+        this.remark = remark;
+    }
+
+    @Override
+    public String toString() {
+        return "Topic{" +
+                "key=" + key +
+                ", label='" + label + '\'' +
+                ", remark='" + remark + '\'' +
+                '}';
+    }
+}